Added Simple Detection Files
Includes simpleDetection.py and the settings.json files. When simpleDetection is called, it will look at the current lidar data that is being gathered and if something is in the lidar's range (currently 30 meters), it will return true, prompting the collision avoidance algorithm to act. Settings.json may have to be tweaked (mainly just the horizontal and vertical settings)